// packages/plugin-calendar-ui/src/settings/types.ts
export type openingHour = {
  days: [string];
  start: string;
  end: string;
};

export type additionalField = {
  label: string;
  name: string;
  required: boolean;
  type: string;
};

export type SchedulePageMutationVariables = {
  _id?: string;
  name: string;
  slug: string;
  timezone: string;
  calendarIds: [string];
  event: {
    title: string;
    location: string;
    duration: number;
  };
  appearance?: {
    color: string;
    companyName?: string;
    logo?: string;
    submitText?: string;
    thankYouText?: string;
  };
  booking?: {
    openingHours?: openingHour[];
    additionalFields?: additionalField[];
    cancellationPolicy?: string;
    confirmationMethod?: string;
    minBookingNotice?: number;
    availableDaysInFuture?: number;
    minBuffer?: number;
    minCancellationNotice?: number;
  };
};
// ...
export type IPage = {
  _id: string;
  name: string;
  slug: string;
  createdAt: Date;
  modifiedAt: Date;
  accountId: string;
  config: {
    timezone: string;
    pageCalendarIds?: any;
    calendarIds?: string[];
    event: {
      title: string;
      location: string;
      duration: number;
    };
    appearance: {
      color?: string;
      companyName?: string;
      logo?: string;
      submitText?: string;
      thankYouText?: string;
      showAutoschedule?: boolean;
      showNylasBranding?: boolean;
    };
    booking: {
      openingHours: openingHour[];
      additionalFields?: additionalField[];
      cancellationPolicy?: string;
      confirmationMethod?: string;
      minBookingNotice?: number;
      availableDaysInFuture?: number;
      minBuffer?: number;
      minCancellationNotice?: number;
    };
  };
};